[UPDATE] Have a true lazy feature with LazyInMemoryProvider
