+++ src/Symfony/Bundle/FeatureToggleBundle/Resources/config/providers.php
@@ -11,15 +11,14 @@
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
-use Symfony\Bundle\FeatureToggleBundle\Provider\LazyInMemoryProvider;
+use Symfony\Component\FeatureToggle\Provider\LazyInMemoryProvider;
 
 return static function (ContainerConfigurator $container) {
     $services = $container->services();
 
     $services->set('feature_toggle.provider.lazy_in_memory', LazyInMemoryProvider::class)
         ->args([
-//            '$features' => abstract_arg('Defined in FeatureToggleExtension'),
-            '$providerLocator' => abstract_arg('Defined in FeatureToggleExtension'),
+            '$features' => abstract_arg('Defined in FeatureToggleExtension'),
         ])
         ->tag('feature_toggle.feature_provider', ['priority' => 16])
     ;
